requirements.txt
# ── Core UI ───────────────────────────────────────────────────────────────────
chainlit>=1.3.0

# ── LangChain stack ───────────────────────────────────────────────────────────
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-pinecone>=0.2.0
langchain-text-splitters>=0.3.0

# ── Vector DB ─────────────────────────────────────────────────────────────────
pinecone-client>=3.0.0

# ── PDF parsing ───────────────────────────────────────────────────────────────
pypdf>=4.0.0

# ── Environment ───────────────────────────────────────────────────────────────
python-dotenv>=1.0.0